Moved server tests to test package and changed test launcher
