+++ appveyor.yml
@@ -20,7 +20,7 @@ install:
   - "%PYTHON%/Scripts/pip.exe install -r conans/requirements_server.txt"
 
 test_script:
-  - "nosetests --with-coverage"
+  - "nosetests --with-coverage" conans.test
   
 after_test:
   - "codecov"
